add some tests and class method implementations
