+++ requirements.txt
@@ -1,4 +1,4 @@
 pip>=21.1
-numpy==1.21.0
+numpy==1.22.0
 matplotlib==3.4.2
 scipy==1.7.0
