+++ test/__init__.py
@@ -1,0 +1,1 @@
+import os, sys; sys.path.append(os.path.dirname(os.path.realpath('..')))
